Add specific subclasses of WebSocketServlet to create websockets of specific types
