+++ server/src/com/thoughtworks/go/server/websocket/agent/AgentRemoteServlet.java
@@ -16,26 +16,10 @@
 
 package com.thoughtworks.go.server.websocket.agent;
 
-import com.thoughtworks.go.util.SystemEnvironment;
-import org.eclipse.jetty.websocket.servlet.WebSocketServlet;
-import org.eclipse.jetty.websocket.servlet.WebSocketServletFactory;
-import org.springframework.web.context.WebApplicationContext;
-import org.springframework.web.context.support.WebApplicationContextUtils;
-
-import javax.servlet.ServletException;
-
-public class AgentRemoteServlet extends WebSocketServlet {
-    private WebApplicationContext wac;
-
+public class AgentRemoteServlet extends AbstractWebSocketServlet<AgentRemoteSocketCreator> {
     @Override
-    public void init() throws ServletException {
-        this.wac = WebApplicationContextUtils.getWebApplicationContext(getServletContext());
-        super.init();
+    protected Class<AgentRemoteSocketCreator> socketCreater() {
+        return AgentRemoteSocketCreator.class;
     }
 
-    @Override
-    public void configure(WebSocketServletFactory factory) {
-        factory.getPolicy().setIdleTimeout(new SystemEnvironment().getWebsocketMaxIdleTime());
-        factory.setCreator(wac.getBean(AgentRemoteSocketCreator.class));
-    }
 }
